+++ python/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Added support for recording OpenReview IDs for authors.
+- Added support for (redundant) math-mode delimiters within `<tex-math>` tags.
+
 ## [1.2.0] — 2026-05-21
 
 This release refactors internals of the library to improve modification tracking and dynamic updating of relevant attributes when data is modified.  The breaking changes here are mostly relevant if you are modifying data.